requirements.txt
asgiref==3.11.1
Django==6.0.2
django-debug-toolbar==6.2.0
djangorestframework==3.16.1
dotenv==0.9.9
drf-nested-routers==0.95.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
